Creating a new entity resolver which will handle requesting entities
This means we can provide easy methods to lookup a whole lot of users and stuff in responses. This will hook into the entire gateway to intercept responses from microservices before they are handled by the normal HTTP response bit. This will be used to populate responses from microservices so they don't have to be aware of each other