Stripping out existing callbacks, going to generic handlers, new types
The new uems comms lib adds the new event types so a lot of this is now out of date. Updating things to at least build without errors but will need to come back and make sure that everything is working correctly once the event microservice is updated in line with this. Currently going through each service bit by bit and working through the updates