Removing set loaders and fixing resolution
Set loaders are really error prone and I've been making weird changes to make them work so its easier to just remove them. Additionally the entity loaders were causing lots of problems because I forgot about subdependencies so now when you load an entity it will load all subentities for it as well. This will not support need loading however which should be fixed in a future release if we reach the point of needing it