Forgot req.query values are always strings, breaks the type system
Right now I'm asserting that they are numbers and booleans but that doesn't work at all so adding a new system to coerce them into the types that they need to be and failing if they can't be coerced. This actually works. This will have some knock on effects in the next commit