Adding tests for handleDefaultResponse and fixing bugs
Previously if the transformer is not asynchronous and throws an error it would crash things but now its wrapped in a try catch to ensure that it is handled properly. Additionally the function now returns a promise which will resolve when its done. This is required by the test to be able to schedule things properly but should also allow a bit more flexibility rather than just hiding the promise