Adding tests for the entity resolver
This one is not as complete as the others because I'm not going to go into testing the individual resolving functions. As long as the base system works I'm happy. This generally tests that the mechanisms of the entity resolver work but there are a few changes to make it nice. There is now a stop function on the entity resolver which will cancel the interval. This means that jest can actually stop as it has no open handles/processes. Made resolve public for the testing as well which isn't a great thing to do but meh. It works.