Adding initial set of role mappings
These roles are required to perform specific actions. A lot of exceptions have been added right now which need handling beyond the simple route level as they depend on things like user IDs and the returned data. These will be implemented in another way but this is the start of the permission systems