Creating a set of lookups
Using this new ID system, I'm adding a system to look up values. This is basically the resolver system but I'm looking to migrate away from that so this will be the new system. Described more in flows of lookups rather than looking up every value required as they come in.